Initial support for bridging Matrix bans to IRC
The bridge used to *not* bridge Matrix bans to IRC. As such, when
someone banned an IRC user from Matrix, it would only prevent Matrix
users from seeing messages coming from said user, but would not ban that
user from IRC at all. This resulted in Matrix channel moderators being
confused when other IRC users are reporting spam that simply isn't
bridged at all as a result.

This commit adds support for bridging Matrix bans to IRC. Currently, it
just bans on IRC based on the IRC user's nickname, but this could change
in the future, and most importantly is better than not bridging the
ban at all.

Signed-off-by: Ferass El Hafidi <vitali64pmemail@protonmail.com>